fix(ui): compose AI operator into existing action-center shell slot
